requirements.txt
acme==1.21.0
apturl==0.5.2
asgiref==3.5.2
blinker==1.4
Brlapi==0.8.3
certbot==1.21.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.2
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-filter==22.1
    django-phonenumber-field==7.0.0
djangorestframework==3.14.0
drf-yasg==1.21.4
gunicorn==20.1.0
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
josepy==1.10.0
keyboard==0.13.5
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==2.1.1
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
parsedatetime==2.6
pexpect==4.8.0
phonenumbers==8.12.57
Pillow==9.0.1
protobuf==3.12.4
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.3
systemd-python==234
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
